כל מיני שינויים בצד מנהל

+++ src/app/components/administor-components/edit-products/edit-products.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { ProductsService } from '../../../services/products.service';
+import { AuthService } from '../../../services/auth.service';
 @Component({
   selector: 'app-edit-products',
   templateUrl: './edit-products.component.html',
@@ -7,8 +8,8 @@ import { ProductsService } from '../../../services/products.service';
 })
 export class EditProductsComponent implements OnInit {
   public showAdd: boolean = true;
-  arr = ["AKC6338 מיטה מלבנית משובצת.JPG", "AKC4600 מיטת אביב מלבנית.JPG","מיטת-זמש-אורטופדית-250x150.jpg"]
-  constructor(public productsService: ProductsService) {
+  arr = ["AKC6338 מיטה מלבנית משובצת.JPG", "AKC4600 מיטת אביב מלבנית.JPG", "מיטת-זמש-אורטופדית-250x150.jpg"]
+  constructor(public productsService: ProductsService, private authService: AuthService) {
   
     }
   ngOnInit() {
@@ -21,6 +22,7 @@ export class EditProductsComponent implements OnInit {
 
   }
   delete(product) {
+    if (this.authService.ifDel==true)
     this.productsService.deleteProduct(product._id);
   }
 }
